Rename poly_text.c
