+++ include/poly_text.h
