+++ main.lsf
@@ -182,7 +182,7 @@ Overlay OVY_28
 Overlay OVY_29
 {
 	After OVY_10
-	Object src/overlay_29_022DC240.o
+	Object src/poly_text.o
 	Object asm/overlay_29_022DC2B8.o
 	Object src/dg_022E0354.o
 	Object asm/overlay_29_022E0378.o
